Note -latomic needed when using -mcx16
The ABA example relies on 16-byte atomic compare-and-swap operations,
which require the `-mcx16` compiler flag on x86-64 to enable the
corresponding instructions.

However, due to current limitations in GCC, 16-byte atomic operations
(e.g., `__atomic_load_16` and `__atomic_compare_exchange_16`) are
implemented via calls to `libatomic` on many platforms. This results
in linker errors unless `-latomic` is explicitly added.

To avoid confusion and ensure portability, this patch adds a footnote
to the LaTeX document explaining that `-latomic` must be linked when
using `-mcx16`. For more details, see:
https://gcc.gnu.org/bugzilla/show_bug.cgi?id=104688